FIXME: net: ipa: Add support for IPA v2.x endpoints
IPA v2.x endpoints are the same as the endpoints on later versions. The
only big change was the addition of the "skip_config" flag. The only
other change is the backlog limit, which is a fixed number for IPA v2.6L

Signed-off-by: Sireesh Kodali <sireeshkodali1@gmail.com>
[Yassine: Don't use aggr_force_close_fmask(), return false in
ipa_endpoint_aggr_active for v2.*, don't set ROUTE_DEF_RETAIN_HDR
for v2.*]
Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>